+++ src/Rasa.Game/Structures/Manifestation.cs
@@ -38,8 +38,8 @@ namespace Rasa.Structures
         public Dictionary<int, MissionLog> Missions { get; set; } = new Dictionary<int, MissionLog>();
         public DateTime LoginTime { get; set; }
         public List<int> Logos = new List<int>();
-        public uint TargetEntityId { get; set; }
-        public uint TrackingTargetEntityId { get; set; }
+        public ulong TargetEntityId { get; set; }
+        public ulong TrackingTargetEntityId { get; set; }
         public byte ActiveWeapon { get; set; }
         public List<uint> GainedWaypoints = new List<uint>();
